requirements.txt
drf-nested-routers==0.9.0
gunicorn==19.1.1
six==1.8.0
static3==0.5.1
aniso8601==1.0.0
Babel==1.3
blinker==1.3
boto==2.38.0
certifi==2015.4.28
coverage==3.7.1
decorator==3.4.0
defusedxml==0.4.1
Django==1.8.2
django-appconf==1.0.1
django-compressor==1.5
django-filter==0.10.0
djangorestframework==3.1.3
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
marshmallow==1.2.2
mysqlclient==1.3.6
oauthlib==0.7.2
requests==2.7.0
thrift==0.9.2
